+++ air-script/tests/selectors/mod.rs
@@ -1,9 +1,15 @@
 #[rustfmt::skip]
 #[allow(clippy::all)]
+mod selectors;
+#[rustfmt::skip]
+#[allow(clippy::all)]
 mod selectors_combine_simple;
 #[rustfmt::skip]
 #[allow(clippy::all)]
 mod selectors_combine_complex;
+#[rustfmt::skip]
+#[allow(clippy::all)]
+mod selectors_combine_with_list_comprehensions;
 #[rustfmt::skip]
 #[allow(clippy::all)]
 mod selectors_with_evaluators;
@@ -19,6 +25,10 @@ mod selectors_combine_complex_plonky3;
 #[allow(clippy::all)]
 #[allow(unused_imports)]
 mod selectors_with_evaluators_plonky3;
+#[rustfmt::skip]
+#[allow(clippy::all)]
+#[allow(unused_imports)]
+mod selectors_combine_with_list_comprehensions_plonky3;
 
 mod test_air_plonky3;
 mod test_air_winterfell;