+++ CHANGELOG.md
@@ -3,11 +3,12 @@
 ## 0.5.0 (TBD)
 
 - Incremented MSRV to 1.89.
-- Add a constant propagation pass after other mir passes (#439).
 - Add a pass for common subexpression elimination in the constraint graph (#419).
+- Reorder unrolling passes to unroll list comprehensions before match statements (#431).
 - Refactored the unrolling pass in MIR (#434).
-- Fix regressions on MIR and list_comprehensions (#449).
 - Update documentation and tests thereof (#437).
+- Add a constant propagation pass after other mir passes (#439).
+- Fix regressions on MIR and list_comprehensions (#449).
 - Add Plonky3 codegen backend (#461).
 
 ## 0.4.0 (2025-06-20)